tests: fix mode is single but not correct behavior
