+++ tests/BaseSelect.test.tsx
@@ -123,51 +123,4 @@ describe('BaseSelect', () => {
 
     expect(container.querySelector('.rc-select-dropdown-placement-fallback')).toBeTruthy();
   });
-
-  describe("Testing BaseSelect component's onContainerBlur params", () => {
-    it('mode with null, onBlur source is blur', () => {
-      const onSearch = jest.fn();
-      const { container } = render(
-        <BaseSelect
-          prefixCls="rc-select"
-          id="test"
-          displayValues={[]}
-          onDisplayValuesChange={() => {}}
-          searchValue="1"
-          showSearch
-          onSearch={onSearch}
-          OptionList={OptionList}
-          emptyOptions
-        />,
-      );
-      expect(container.querySelector('div.rc-select')).toBeTruthy();
-      fireEvent.change(container.querySelector('input'), { target: { value: '2' } });
-      expect(onSearch).toHaveBeenCalledWith('2', { source: 'typing' });
-      fireEvent.blur(container.querySelector('div.rc-select'));
-      expect(onSearch).toHaveBeenCalledWith('', { source: 'blur' });
-    });
-
-    it('mode with multiple, onBlur source is blur', () => {
-      const onSearch = jest.fn();
-      const { container } = render(
-        <BaseSelect
-          prefixCls="rc-select"
-          mode="multiple"
-          id="test"
-          displayValues={[]}
-          onDisplayValuesChange={() => {}}
-          searchValue="1"
-          showSearch
-          onSearch={onSearch}
-          OptionList={OptionList}
-          emptyOptions
-        />,
-      );
-      expect(container.querySelector('div.rc-select')).toBeTruthy();
-      fireEvent.change(container.querySelector('input'), { target: { value: '2' } });
-      expect(onSearch).toHaveBeenCalledWith('2', { source: 'typing' });
-      fireEvent.blur(container.querySelector('div.rc-select'));
-      expect(onSearch).toHaveBeenCalledWith('', { source: 'blur' });
-    });
-  });
 });
